Polyfit generalisation (#46)
Lightcurve fits using polynomials is highly unstable due to different
results depending on polynomial degree.

![Polyfit](https://user-images.githubusercontent.com/70367168/215088537-54c96a36-c1b5-492e-838b-2d834b8bbd78.png)
1. The day of maximum light varies by a couple of days depending on the
degree
2. Towards the end of the lightcurve/ the end of the simulation, the fit
fails. This becomes critical when calculating delta_m40. Since the day
of maximum light varies strongly, it is also not an option (in this
particular case) to exclude the last points to achieve a better fit.

### Proposed solution:
Default to lightcurve fitting using a Gaussian Process regression using
`george` including a hyperparameter optimization. Runtime overhead is
barely noticeable and this way it is not needed to make a choice for a
particular degree of polynomial. The only choice here is the choice of
kernel (Matern5/2; this choice should have a rather limited effect). The
plot below shows the result:

![Maternfit](https://user-images.githubusercontent.com/70367168/215089540-6195ef1b-90f8-4463-ba36-9335d12d8afe.png)

*(Note: Polynomial fits are only plotted for reference here. All fits
have been done on the same data)*

In case george is not found, it will fall back to the old polyfit.