Add cell merging to e2e mapping (#375)
Add an option to the script which merges a square number of cells after
the main script is run. Main purpose: If some potentially critical mass
fractions are not properly mapped by eg. 25x50 cells, the consistency in
mass fractions can be improved by mapping to 100x200 cells first and
then merging cells in the ARTIS model to downgrade to a 25x50 model
again, now with more consistent mass fractions.